refactor project add dialog
