+++ ui/src/features/Dashboard/components/MainSection/components/index.ts
@@ -1,3 +1,4 @@
 export { ProjectDeletionDialog } from "./ProjectDeletionDialog";
 export { ProjectEditDialog } from "./ProjectEditDialog";
+export { ProjectAddDialog } from "./ProjectAddDialog";
 export { ProjectCard } from "./ProjectCard";
