+++ ui/src/features/Dialog/components/Content.tsx
@@ -5,7 +5,6 @@ import {
 import { DialogType } from "@flow/stores";
 
 import useInstructions from "./instructions/useInstructions";
-import useProject from "./project/useProject";
 import useSearches from "./searches/useSearches";
 import useSettings from "./settings/useSettings";
 import useWorkspace from "./workspace/useWorkspace";
@@ -28,7 +27,6 @@ const DialogContent: React.FC<Props> = ({ tab, position, onTabChange }) => {
   const settings = useSettings();
   const instructions = useInstructions();
   const addWorkspace = useWorkspace();
-  const addProject = useProject();
 
   // TODO: Isn't this very hackish?
   const content = tab?.includes("search")
@@ -39,9 +37,7 @@ const DialogContent: React.FC<Props> = ({ tab, position, onTabChange }) => {
         ? instructions
         : tab?.includes("workspace")
           ? addWorkspace
-          : tab?.includes("project")
-            ? addProject
-            : null;
+          : null;
 
   const disableClickAway = tab.includes("settings");
