--- ui/src/features/Dialog/components/project/index.ts
@@ -1,1 +1,0 @@
-export * from "./AddProject";
